Working. Next is to refactor the model so that it returns transformed points for the view to draw.

+++ src/main/java/org/example/GlobalVars.java
@@ -5,7 +5,7 @@ public class GlobalVars {
     public static final int LEFT_ISO = 0;
     public static final int RIGHT_ISO = 2;
     public static int mode = 1;
-    public static float ySkew = 0.5f;
+    public static double ySkew = 0.5f;
 
     private GlobalVars() {
     }
